shapes

+++ projects/zloshnik/src/src/text.h
@@ -9,8 +9,8 @@ class Text {
 private:
     uint8_t text_angle;
 public:
-    void Char(uint8_t c, int ox, int oy, uint8_t scale);
-    void Str(const char *s, uint8_t scale, uint8_t angle);
+    void Char(uint8_t c, int ox, int oy, int16_t scale);
+    void Str(const char *s, int16_t scale, uint8_t angle);
 };
 
 extern Text text;
